+++ src/linker/utilities/spark_utils.py
@@ -1,0 +1,11 @@
+
+
+def build_cluster():
+    """Builds a Spark cluster.
+
+    Returns:
+        spark_master_url: Spark master URL.
+    """
+    # TODO implement
+    ...
+
